+++ lib/main.dart
@@ -22,12 +22,12 @@ import 'package:flutter/foundation.dart';
 import 'package:illinois/service/AppDateTime.dart';
 import 'package:illinois/service/Auth2.dart';
 import 'package:illinois/service/Canvas.dart';
+import 'package:illinois/service/CheckList.dart';
 import 'package:illinois/service/DeepLink.dart';
 import 'package:illinois/service/DeviceCalendar.dart';
 import 'package:illinois/service/Dinings.dart';
 import 'package:illinois/service/FirebaseMessaging.dart';
 import 'package:illinois/service/FlexUI.dart';
-import 'package:illinois/service/Gies.dart';
 import 'package:illinois/service/Guide.dart';
 import 'package:illinois/service/IlliniCash.dart';
 import 'package:illinois/service/LiveStats.dart';
@@ -125,7 +125,8 @@ void main() async {
     DeviceCalendar(),
     Events(),
     Groups(),
-    Gies(),
+    CheckList("gies"),
+    CheckList("uiuc_student"),
     Canvas(),
     Rewards(),
     OnCampus(),