+++ CHANGELOG.md
@@ -5,10 +5,14 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## Unreleased
+## [4.0.17] - 2022-05-27
 ### Changed
 - Buttons for "On Campus" [#1664](https://github.com/rokwire/illinois-app/issues/1664).
 - Fix typo in HomeRadioWidget [#1666](https://github.com/rokwire/illinois-app/issues/1666).
 - Change comply message when creating a Group [#1668](https://github.com/rokwire/illinois-app/issues/1668).
+- Implemented drag and drop capability in HomePanel [#1629](https://github.com/rokwire/illinois-app/issues/1629).
+### Added
+- Regular Student onboarding checklist (in progress) [#1671](https://github.com/rokwire/illinois-app/issues/1671).
 
 ## [4.0.16] - 2022-05-25
 ### Changed